requirements.txt
amqp==5.1.1
asgiref==3.6.0
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
class-registry==2.1.2
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.4
Django==3.2.16
django-factory-boy==1.0.0
django-filter==22.1
django-psycopg2-extension==0.1.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
factory-boy==3.2.1
Faker==15.3.4
filters==1.3.2
filters-django==1.0.5
inflection==0.5.1
jsonschema==4.17.3
kombu==5.2.4
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycparser==2.21
pyrsistent==0.19.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.7
PyYAML==6.0
regex==2022.10.31
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.5
